refactored error boundary so it doesnt use context
